Fix Test Script

+++ lib/sow-payload-parser.ts
@@ -4,7 +4,6 @@
 * See the accompanying LICENSE file for terms.
 */
 // 11:17 PM 5/5/2020
-import { IncomingHttpHeaders } from 'http';
 import { IRequest } from './sow-server-core';
 import * as _fs from 'fs';
 import * as _path from 'path';
@@ -63,9 +62,10 @@ const getLine = ( req: IRequest, data: Buffer[] ): string => {
             case '\r':
                 outstr += str; data.push( c );
                 c = req.read( 1 );
-                if ( !c ) return outstr;
-                outstr += c.toString();// assume \n
-                data.push( c );
+                if ( c ) {
+                    outstr += c.toString();// assume \n
+                    data.push( c );
+                }
                 return outstr;
             default:
                 outstr += str;
@@ -73,12 +73,7 @@ const getLine = ( req: IRequest, data: Buffer[] ): string => {
         }
     }
 }
-const getHeader = ( headers: IncomingHttpHeaders, key: string ): string => {
-    const result: string | string[] | undefined = headers[key];
-    return typeof ( result ) === "string" ? result.toString() : "";
-}
 const incomingContentType: {
-    [key: string]: any;
     URL_ENCODE: string;
     APP_JSON: string;
     MULTIPART: string;
@@ -203,7 +198,8 @@ export class PostedFileInfo implements IPostedFileInfo {
         if ( this._isDisposed ) return;
         this._isDisposed = true;
         if ( !this._isMoved && this._tempFile ) {
-            _fs.unlinkSync( this._tempFile );
+            if ( _fs.existsSync( this._tempFile ) )
+                _fs.unlinkSync( this._tempFile );
         }
         delete this._fcontentDisposition;
         delete this._fname;
@@ -324,7 +320,7 @@ class PayloadDataParser {
         }
     }
     private endCurrentStream( exit: boolean ): void {
-        if ( this._writeStream && this._isStart === true ) {
+        if ( this._isStart === true && this._writeStream ) {
             this._writeStream.end();
             if ( this._postedFile ) {
                 if ( exit === false ) {
@@ -367,8 +363,8 @@ export class PayloadParser implements IPayloadParser {
         tempDir: string
     ) {
         this._isDisposed = false; this._isConnectionActive = true;
-        this._contentType = getHeader( req.headers, "content-type" );
-        this._contentLength = ToNumber( getHeader( req.headers, "content-length" ) );
+        this._contentType = req.get( "content-type" ) || "";
+        this._contentLength = ToNumber( req.get( "content-length" ) || 0 );
         if ( this._contentType.indexOf( incomingContentType.MULTIPART ) > -1 ) {
             this._contentTypeEnum = ContentType.MULTIPART;
         } else if ( this._contentType.indexOf( incomingContentType.URL_ENCODE ) > -1 && this._contentType === incomingContentType.URL_ENCODE ) {
